+++ CHANGELOG.md
@@ -8,6 +8,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [In Development] - Unreleased
 
+### Added
+
+- Default forum settings to the admin backend
+
+### Changed
+
+- `Setting` model is now a singleton to prevent multiple settings of the same type
+
 
 ## [1.10.0] - 2022-08-03